move AsciiString stuff to strings.py
