+++ ofrak_core/ofrak/core/strings.py
@@ -1,11 +1,20 @@
+import asyncio
 import re
 from dataclasses import dataclass
 
+from ofrak.component.analyzer import Analyzer
 from ofrak.component.modifier import Modifier, ModifierError
+from ofrak.component.unpacker import Unpacker
 from ofrak.core.binary import BinaryPatchConfig, BinaryPatchModifier, GenericText, GenericBinary
+from ofrak.core.code_region import CodeRegion
 from ofrak.core.filesystem import File
+from ofrak.core.program_section import ProgramSection
 from ofrak.model.component_model import ComponentConfig
+from ofrak.model.resource_model import index
+from ofrak.model.viewable_tag_model import AttributesType
 from ofrak.resource import Resource
+from ofrak.resource_view import ResourceView
+from ofrak_type import Range
 
 
 @dataclass
@@ -77,3 +86,58 @@ class StringFindReplaceModifier(Modifier[StringFindReplaceConfig]):
         offsets = [m.start() for m in re.finditer(to_find, original_data)]
         for offset in offsets:
             await resource.run(BinaryPatchModifier, BinaryPatchConfig(offset, replace_with))
+
+
+@dataclass
+class AsciiString(ResourceView):
+    text: str
+
+    @index
+    def Text(self) -> str:
+        return self.text
+
+    @classmethod
+    def caption(cls, all_attributes) -> str:
+        try:
+            s = all_attributes[AttributesType[AsciiString]]
+            return f"string: '{s.text}'"
+        except KeyError:
+            return super().caption(all_attributes)
+
+
+class AsciiStringAnalyzer(Analyzer[None, AsciiString]):
+    targets = (AsciiString,)
+    outputs = (AsciiString,)
+
+    async def analyze(self, resource: Resource, config: None) -> AsciiString:
+        raw_without_null_byte = (await resource.get_data())[:-1]
+        return AsciiString(raw_without_null_byte.decode("ascii"))
+
+
+class StringsUnpacker(Unpacker[None]):
+    targets = (ProgramSection,)  # TODO: Other reasonable targets?
+    children = (AsciiString,)
+
+    async def unpack(self, resource: Resource, config: None) -> None:
+        if resource.get_data_id() is None:
+            return
+        if resource.has_tag(CodeRegion):
+            # code is less likely to have strings so more likely to have false positives
+            min_length = 8
+        else:
+            min_length = 2
+
+        # match sequences of printable characters of at least `min_length` ending with null byte
+        # printable characters defined as: ASCII between ' ' and '~', tab, newline, carriage return
+        pattern = rb"([ -~,\n,\t,\r]{%d,})\x00" % min_length
+
+        data = await resource.get_data()
+
+        children = [
+            resource.create_child_from_view(
+                AsciiString(m.group(0).decode("ascii")), data_range=Range(m.start(), m.end())
+            )
+            for m in re.finditer(pattern, data)
+        ]
+
+        await asyncio.gather(*children)
