+++ ofrak_core/ofrak/core/strings_analyzer.py
@@ -1,24 +1,12 @@
 import asyncio
-import re
 import tempfile
 from dataclasses import dataclass
 from typing import Dict, Optional
 
-from ofrak.model.viewable_tag_model import AttributesType
-
-
-from ofrak.core import CodeRegion
-
-from ofrak.core.program_section import ProgramSection
-
-from ofrak.component.unpacker import Unpacker
-
 from ofrak.component.analyzer import Analyzer
 from ofrak.resource import Resource
 from ofrak.model.component_model import ComponentConfig
-from ofrak.model.resource_model import ResourceAttributes, index
-from ofrak.resource_view import ResourceView
-from ofrak_type import Range
+from ofrak.model.resource_model import ResourceAttributes
 
 
 @dataclass
@@ -69,58 +57,3 @@ class StringsAnalyzer(Analyzer[Optional[StringsAnalyzerConfig], StringsAttribute
             await proc.wait()
 
         return StringsAttributes(strings)
-
-
-@dataclass
-class AsciiString(ResourceView):
-    text: str
-
-    @index
-    def Text(self) -> str:
-        return self.text
-
-    @classmethod
-    def caption(cls, all_attributes) -> str:
-        try:
-            s = all_attributes[AttributesType[AsciiString]]
-            return f"string: '{s.text}'"
-        except KeyError:
-            return super().caption(all_attributes)
-
-
-class AsciiStringAnalyzer(Analyzer[None, AsciiString]):
-    targets = (AsciiString,)
-    outputs = (AsciiString,)
-
-    async def analyze(self, resource: Resource, config: None) -> AsciiString:
-        raw_without_null_byte = (await resource.get_data())[:-1]
-        return AsciiString(raw_without_null_byte.decode("ascii"))
-
-
-class StringsUnpacker(Unpacker[None]):
-    targets = (ProgramSection,)  # TODO: Other reasonable targets?
-    children = (AsciiString,)
-
-    async def unpack(self, resource: Resource, config: None) -> None:
-        if resource.get_data_id() is None:
-            return
-        if resource.has_tag(CodeRegion):
-            # code is less likely to have strings so more likely to have false positives
-            min_length = 8
-        else:
-            min_length = 2
-
-        # match sequences of printable characters of at least `min_length` ending with null byte
-        # printable characters defined as: ASCII between ' ' and '~', tab, newline, carriage return
-        pattern = rb"([ -~,\n,\t,\r]{%d,})\x00" % min_length
-
-        data = await resource.get_data()
-
-        children = [
-            resource.create_child_from_view(
-                AsciiString(m.group(0).decode("ascii")), data_range=Range(m.start(), m.end())
-            )
-            for m in re.finditer(pattern, data)
-        ]
-
-        await asyncio.gather(*children)
